Fix numberformat
